unit tests refactored and now running during mvn install
