+++ .gitignore
@@ -1,15 +1,40 @@
-/target/
-/.DS_Store
+# IntelliJ
+.idea/
+*.eml
+*.iml
+*.ipr
+*.iws
+
+# Eclipse
 .classpath
 .project
-.settings
+.metadata
+.settings/
+local.properties
+
+# Maven, etc.
+out/
+target/
+
+# General
+*.pyc
+*~
+*.bak
+*.sw[o,p]
+*.tmp
+.DS_Store
+Thumbs.db
 
 *.class
+
 # Mobile Tools for Java (J2ME)
 .mtj.tmp/
+
 # Package Files #
 *.jar
 *.war
 *.ear
+
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+
